+++ src/Threebox.js
@@ -113,6 +113,10 @@ Threebox.prototype = {
 			//[jscastro] if multiLayer, create a by default layer in the map, so tb.update won't be needed in client side to avoid duplicating calls to render
 			if (this.tb.options.multiLayer) this.addLayer({ id: "threebox_layer", type: 'custom', renderingMode: '3d', map: this, onAdd: function (map, gl) { }, render: function (gl, matrix) { this.map.tb.update(); } })
 
+			this.once('idle', () => {
+				this.tb.setObjectsScale();
+			});
+
 			if (this.tb.options.sky) {
 				this.tb.sky = true;
 			}
@@ -489,7 +493,7 @@ Threebox.prototype = {
 
 			this.onZoom = function (e) {
 				this.tb.zoomLayers.forEach((l) => { this.tb.toggleLayer(l); });
-				this.tb.world.children.filter(o => (o.fixedZoom != null)).forEach((o) => { o.setObjectScale(this.transform.scale); });
+				this.tb.setObjectsScale();
 			}
 
 			let ctrlDown = false;
@@ -816,12 +820,7 @@ Threebox.prototype = {
 		this.map.setLayoutProperty(layerId, name, value);
 		if (value !== null && value !== undefined) {
 			if (name === 'visibility') {
-				this.world.children.forEach(function (obj) {
-					if (obj.layer === layerId) {
-						obj.visibility = value;
-					}
-				});
-				return;
+				this.world.children.filter(o => (o.layer === layerId)).forEach((o) => { o.visibility = value });
 			}
 		}
 	},
@@ -857,6 +856,11 @@ Threebox.prototype = {
 				}
 			});
 		}
+	},
+
+	//[jscastro] method to set globally all the objects that are fixedScale
+	setObjectsScale: function () {
+		this.world.children.filter(o => (o.fixedZoom != null)).forEach((o) => { o.setObjectScale(this.map.transform.scale); });
 	},
 
 	//[jscastro] mapbox setStyle removes all the layers, including custom layers, so tb.world must be cleaned up too